- Getting month from body - Chaning rabbits count to long - Friendly Exception for overflow and general exception

+++ RabbitsLifeCycle.Web.Api/Models/RabbitBirthRegistration.cs
@@ -2,7 +2,7 @@
 {
     public class RabbitBirthRegistration
     {
-        public RabbitBirthRegistration(int month, int bornRabbits)
+        public RabbitBirthRegistration(int month, long bornRabbits)
         {
             Month = month;
             BornRabbits = bornRabbits;
@@ -10,7 +10,7 @@
 
         public int Month { get; set; }
 
-        public int BornRabbits { get; set; }
+        public long BornRabbits { get; set; }
 
         public int MatureMonth => Month + 3;
     }
